plugin: CheckstyleExecutor: extracted from AbstractCheckMojo
