+++ plugin/src/main/java/net/kemitix/checkstyle/ruleset/plugin/CheckstyleExecutor.java
@@ -1,0 +1,53 @@
+/**
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2017 Paul Campbell
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software
+ * and associated documentation files (the "Software"), to deal in the Software without restriction,
+ * including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
+ * and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies
+ * or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE
+ * AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+package net.kemitix.checkstyle.ruleset.plugin;
+
+import org.apache.maven.artifact.repository.ArtifactRepository;
+import org.apache.maven.execution.MavenSession;
+import org.apache.maven.plugin.BuildPluginManager;
+import org.apache.maven.plugin.MojoExecutionException;
+import org.apache.maven.plugin.MojoFailureException;
+import org.apache.maven.project.MavenProject;
+
+/**
+ * Executes the Checkstyle plugin.
+ *
+ * @author Paul Campbell (pcampbell@kemitix.net)
+ */
+interface CheckstyleExecutor {
+
+    /**
+     * Execute Checkstyle Check.
+     *
+     * @param mavenProject       The Maven Project
+     * @param mavenSession       The Maven Session
+     * @param artifactRepository The Artifact Repository
+     * @param pluginManager      The Maven Plugin Manager
+     *
+     * @throws MojoExecutionException on execution error
+     * @throws MojoFailureException   on execution failure
+     */
+    void performCheck(
+            final MavenProject mavenProject, final MavenSession mavenSession,
+            final ArtifactRepository artifactRepository, final BuildPluginManager pluginManager
+                     ) throws MojoExecutionException, MojoFailureException;
+}
